fix warning
Moved make_hir to toplevel
